requirements.txt
gymnasium>=0.29.0
torch>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0
seaborn>=0.12.0
pandas>=2.0.0

